requirements.txt
absl-py==2.1.0
albumentations==1.4.0
annotated-types==0.7.0
anyio==4.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
blis==0.7.11
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
colorama==0.4.6
comm==0.2.2
confection==0.1.5
contourpy==1.3.0
cycler==0.12.1
cymem==2.0.8
debugpy==1.8.6
decorator==5.1.1
defusedxml==0.7.1
exceptiongroup==1.2.2
executing==2.1.0
fastai==2.7.15
fastcore==1.5.55
fastdownload==0.0.7
fastjsonschema==2.20.0
fastprogress==1.0.3
filelock==3.16.1
flask==3.0.3
fonttools==4.54.1
fsspec==2024.9.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
grpcio==1.66.1
huggingface-hub==0.25.1
idna==3.10
imageio==2.35.1
importlib-metadata==8.5.0
importlib-resources==6.4.5
ipykernel==6.29.5
ipython==8.18.1
ipython-genutils==0.2.0
ipywidgets==7.7.1
itsdangerous==2.2.0
jedi==0.19.1
jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-client==8.6.3
jupyter-core==5.7.2
jupyter-events==0.10.0
jupyter-server==2.14.2
jupyter-server-terminals==0.5.3
jupyterlab-pygments==0.3.0
jupyterlab-widgets==3.0.13
kiwisolver==1.4.7
langcodes==3.4.1
language-data==1.2.0
lazy-loader==0.4
marisa-trie==1.2.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
murmurhash==1.0.10
nbclassic==1.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
notebook==6.5.4
notebook-shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
opencv-contrib-python==4.10.0.82
opencv-python==4.5.5.64
opencv-python-headless==4.10.0.84
overrides==7.7.0
packaging==24.1
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.4
Pillow==9.5.0
platformdirs==4.3.6
preshed==3.0.9
prometheus-client==0.21.0
prompt-toolkit==3.0.48
protobuf==4.23.4
psutil==6.0.0
pure-eval==0.2.3
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic-core==2.23.4
pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.2
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.2
pyzmq==26.2.0
qudida==0.0.4
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.8.1
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.2.2
scipy==1.11.4
seaborn==0.12.2
Send2Trash==1.8.3
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
soupsieve==2.6
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
sympy==1.13.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
terminado==0.18.1
thinc==8.2.5
threadpoolctl==3.5.0
tifffile==2024.8.30
tinycss2==1.3.0
tk==0.1.0
tokenizers==0.19.1
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.41.2
typer==0.12.5
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
websocket-client==1.8.0
werkzeug==3.0.4
widgetsnbextension==3.6.9
wrapt==1.16.0
zipp==3.20.2
